+++ sme_terceirizadas/dados_comuns/permissions.py
@@ -1289,15 +1289,3 @@ class PermissaoParaAnalisarFichaTecnica(BasePermission):
             and usuario.vinculo_atual
             and usuario.vinculo_atual.perfil.nome in self.PERFIS_PERMITIDOS
         )
-
-
-class PermissaoParaVisualizarQuestoesConferencia(BasePermission):
-    PERFIS_PERMITIDOS = [DILOG_QUALIDADE]
-
-    def has_permission(self, request, view):
-        usuario = request.user
-        return (
-            not usuario.is_anonymous
-            and usuario.vinculo_atual
-            and usuario.vinculo_atual.perfil.nome in self.PERFIS_PERMITIDOS
-        )
